+++ src/index.tsx
@@ -4,8 +4,9 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { BrowserRouter } from 'react-router-dom';
 
-import { CartContextProvider } from './stores/CartContext/CartContextProvider';
 import { Router } from './router';
+import { CartContextProvider } from './stores/CartContext';
+import { OrderContextProvider } from './stores/OrderContext';
 
 if (process.env.NODE_ENV === 'development') {
   await import('./mocks/browser').then(({ worker }) => worker.start());
@@ -16,7 +17,9 @@ root.render(
   <React.StrictMode>
     <BrowserRouter>
       <CartContextProvider>
-        <Router />
+        <OrderContextProvider>
+          <Router />
+        </OrderContextProvider>
       </CartContextProvider>
     </BrowserRouter>
   </React.StrictMode>
