+++ src/stores/CartContext/cartStore.ts
@@ -1,65 +1,7 @@
 import { createContext } from 'react';
-import { cloneDeep } from 'lodash-es';
 
-import { CartProductModel, CartProductModelPOJO } from '@/models';
+import { TApiContext, TCartContext, CartStore } from '../commonCartStore';
 
-import { DispatchContext, ReducerReturnType } from '../types';
-
-export type CartStore = { [productId: number]: CartProductModel };
-
-const initialCartStore: CartStore = {};
-
-export function getInitialCardStore() {
-  return cloneDeep(initialCartStore);
-}
-
-type TCartStoreActions = 'add' | 'update' | 'delete';
-
-export function reducer(
-  store: CartStore,
-  action: { type?: TCartStoreActions; payload?: CartProductModel[] | CartProductModelPOJO[] }
-) {
-  const { type, payload } = action;
-  if (!payload) return store;
-  const cartProducts = payload;
-
-  switch (type) {
-    case 'add': {
-      cartProducts.forEach((cartProduct) => {
-        store[cartProduct.product.id] = new CartProductModel(cartProduct);
-      });
-      break;
-    }
-    case 'update': {
-      cartProducts.forEach((cartProduct) => {
-        const targetProduct = store[cartProduct.product.id];
-        if (!targetProduct) return;
-
-        store[cartProduct.product.id] = new CartProductModel(cartProduct);
-      });
-      break;
-    }
-    case 'delete': {
-      cartProducts.forEach((cartProduct) => {
-        const targetProduct = store[cartProduct.product.id];
-        if (!targetProduct) return;
-
-        delete store[cartProduct.product.id];
-      });
-      break;
-    }
-    default: {
-      return getInitialCardStore();
-    }
-  }
-
-  return { ...store };
-}
-
-type TCardReducer = ReducerReturnType<typeof reducer>;
-type Dispatch = TCardReducer[1];
-type TCartContext = CartStore | null;
-type TApiContext = DispatchContext<Dispatch>;
-
+export type TCartStore = CartStore;
 export const CartContext = createContext<TCartContext>(null);
-export const ApiContext = createContext<TApiContext>(null);
+export const CartApiContext = createContext<TApiContext>(null);
