+++ src/stores/CartContext/useCartContextSelector.ts
@@ -1,11 +1,11 @@
 import { useContext } from 'react';
 
-import { ApiContext, CartContext } from './cartStore';
+import { CartApiContext, CartContext } from './cartStore';
 
 export function useCartSelector() {
   return useContext(CartContext);
 }
 
 export function useCartContextApiSelector() {
-  return useContext(ApiContext);
+  return useContext(CartApiContext);
 }
